move TensorProduct into rs
