+++ e3nn/non_linearities/gated_block_parity.py
@@ -3,7 +3,6 @@ import torch
 
 from e3nn import rs
 from e3nn.non_linearities.activation import Activation
-from e3nn.tensor_product import ElementwiseTensorProduct
 
 
 class GatedBlockParity(torch.nn.Module):
@@ -19,7 +18,7 @@ class GatedBlockParity(torch.nn.Module):
         self.act_gates = Activation(Rs_gates, act_gates)
         Rs_gates = self.act_gates.Rs_out
 
-        self.mul = ElementwiseTensorProduct(Rs_nonscalars, Rs_gates)
+        self.mul = rs.ElementwiseTensorProduct(Rs_nonscalars, Rs_gates)
         Rs_nonscalars = self.mul.Rs_out
 
         self.Rs_out = Rs_scalars + Rs_nonscalars
