+++ e3nn/rs.py
@@ -5,7 +5,7 @@ Some functions related to SO3 and his usual representations
 Using ZYZ Euler angles parametrisation
 """
 
-from fractions import gcd
+from math import gcd
 from functools import reduce
 
 import torch
@@ -398,6 +398,91 @@ def tensor_product(input1, input2, output, normalization='component', sorted=Fal
         return Rs_in1, Q
 
 
+class TensorProduct(torch.nn.Module):
+    """
+    Module for tensor_product
+    """
+    def __init__(self, input1, input2, output, normalization='component', sorted=True):
+        super().__init__()
+
+        Rs, mat = tensor_product(input1, input2, output, normalization, sorted)
+
+        self.Rs_in1, self.Rs_in2, self.Rs_out = input1, input2, output
+        if not isinstance(self.Rs_in1, list):
+            self.Rs_in1 = Rs
+        if not isinstance(self.Rs_in2, list):
+            self.Rs_in2 = Rs
+        if not isinstance(self.Rs_out, list):
+            self.Rs_out = Rs
+
+        register_sparse_buffer(self, 'mixing_matrix', mat)
+
+    def __repr__(self):
+        return "{name} ({Rs_in1} x {Rs_in2} -> {Rs_out})".format(
+            name=self.__class__.__name__,
+            Rs_in1=format_Rs(self.Rs_in1),
+            Rs_in2=format_Rs(self.Rs_in2),
+            Rs_out=format_Rs(self.Rs_out),
+        )
+
+    def forward(self, features_1, features_2):
+        '''
+        :param features_1: [..., in1]
+        :param features_2: [..., in2]
+        :return: [..., out]
+        '''
+        d_out = dim(self.Rs_out)
+        d_in1 = dim(self.Rs_in1)
+        d_in2 = dim(self.Rs_in2)
+
+        features = features_1[..., :, None] * features_2[..., None, :]
+
+        size = features.shape[:-2]
+        features = features.reshape(-1, d_in1, d_in2)  # [in1, in2, batch]
+
+        mixing_matrix = get_sparse_buffer(self, "mixing_matrix")  # [out, in1 * in2]
+
+        features = torch.einsum('zij->ijz', features)  # [in1, in2, batch]
+        features = features.reshape(d_in1 * d_in2, features.shape[2])
+        features = mixing_matrix @ features  # [out, batch]
+        return features.T.reshape(*size, d_out)
+
+    def right(self, features_2):
+        '''
+        :param features_2: [..., in2]
+        :return: [..., out, in1]
+        '''
+        d_out = dim(self.Rs_out)
+        d_in1 = dim(self.Rs_in1)
+        d_in2 = dim(self.Rs_in2)
+        size_2 = features_2.shape[:-1]
+        features_2 = features_2.reshape(-1, d_in2)
+
+        mixing_matrix = get_sparse_buffer(self, "mixing_matrix")  # [out, in1 * in2]
+        mixing_matrix = mixing_matrix.sparse_reshape(d_out * d_in1, d_in2)
+        output = mixing_matrix @ features_2.T  # [out * in1, batch]
+        return output.T.reshape(*size_2, d_out, d_in1)
+
+    def left(self, features_1):
+        '''
+        :param features_1: [..., in1]
+        :return: [..., out, in2]
+        '''
+        d_out = dim(self.Rs_out)
+        d_in1 = dim(self.Rs_in1)
+        d_in2 = dim(self.Rs_in2)
+        size_1 = features_1.shape[:-1]
+        features_1 = features_1.reshape(-1, d_in1)
+
+        mixing_matrix = get_sparse_buffer(self, "mixing_matrix")  # [out, in1 * in2]
+        mixing_matrix = mixing_matrix.sparse_reshape(d_out * d_in1, d_in2).t()  # [in2, out * in1]
+        mixing_matrix = mixing_matrix.sparse_reshape(d_in2 * d_out, d_in1)  # [in2 * out, in1]
+        output = mixing_matrix @ features_1.T  # [in2 * out, batch]
+        output = output.reshape(d_in2, d_out, features_1.shape[0])
+        output = torch.einsum('jiz->zij', output)
+        return output.reshape(*size_1, d_out, d_in2)
+
+
 def _tensor_product_in_in(Rs_in1, Rs_in2, selection_rule, normalization, sorted):
     """
     Compute the matrix Q
@@ -675,6 +760,39 @@ def tensor_square(Rs_in, selection_rule=o3.selection_rule, normalization='compon
     return Rs_out, wigner_3j_tensor
 
 
+class TensorSquare(torch.nn.Module):
+    """
+    Module for tensor_square
+    """
+    def __init__(self, Rs_in, selection_rule=o3.selection_rule):
+        super().__init__()
+
+        self.Rs_in = simplify(Rs_in)
+
+        self.Rs_out, mixing_matrix = tensor_square(self.Rs_in, selection_rule, sorted=True)
+        register_sparse_buffer(self, 'mixing_matrix', mixing_matrix)
+
+    def __repr__(self):
+        return "{name} ({Rs_in} ^ 2 -> {Rs_out})".format(
+            name=self.__class__.__name__,
+            Rs_in=format_Rs(self.Rs_in),
+            Rs_out=format_Rs(self.Rs_out),
+        )
+
+    def forward(self, features):
+        '''
+        :param features: [..., channels]
+        '''
+        *size, n = features.size()
+        features = features.reshape(-1, n)
+
+        mixing_matrix = get_sparse_buffer(self, "mixing_matrix")
+
+        features = torch.einsum('zi,zj->ijz', features, features)
+        features = mixing_matrix @ features.reshape(-1, features.shape[2])
+        return features.T.reshape(*size, -1)
+
+
 def elementwise_tensor_product(Rs_in1, Rs_in2, selection_rule=o3.selection_rule, normalization='component'):
     """
     :return: Rs_out, matrix
@@ -752,3 +870,40 @@ def elementwise_tensor_product(Rs_in1, Rs_in2, selection_rule=o3.selection_rule,
         sparse_sizes=(dim(Rs_out), dim(Rs_in1) * dim(Rs_in2)))
 
     return Rs_out, wigner_3j_tensor
+
+
+class ElementwiseTensorProduct(torch.nn.Module):
+    """
+    Module for elementwise_tensor_product
+    """
+    def __init__(self, Rs_in1, Rs_in2, selection_rule=o3.selection_rule):
+        super().__init__()
+
+        self.Rs_in1 = simplify(Rs_in1)
+        self.Rs_in2 = simplify(Rs_in2)
+        assert sum(mul for mul, _, _ in self.Rs_in1) == sum(mul for mul, _, _ in self.Rs_in2)
+
+        self.Rs_out, mixing_matrix = elementwise_tensor_product(self.Rs_in1, self.Rs_in2, selection_rule)
+        register_sparse_buffer(self, "mixing_matrix", mixing_matrix)
+
+    def forward(self, features_1, features_2):
+        '''
+        :param features_1: [..., in1]
+        :param features_2: [..., in2]
+        :return: [..., out]
+        '''
+        d_out = dim(self.Rs_out)
+        d_in1 = dim(self.Rs_in1)
+        d_in2 = dim(self.Rs_in2)
+
+        features = features_1[..., :, None] * features_2[..., None, :]
+
+        size = features.shape[:-2]
+        features = features.reshape(-1, d_in1, d_in2)  # [in1, in2, batch]
+
+        mixing_matrix = get_sparse_buffer(self, "mixing_matrix")  # [out, in1 * in2]
+
+        features = torch.einsum('zij->ijz', features)  # [in1, in2, batch]
+        features = features.reshape(d_in1 * d_in2, features.shape[2])
+        features = mixing_matrix @ features  # [out, batch]
+        return features.T.reshape(*size, d_out)
