+++ tests/tensor_product_test.py
@@ -1,107 +1,9 @@
 # pylint: disable=not-callable, no-member, invalid-name, line-too-long, wildcard-import, unused-wildcard-import, missing-docstring
 import unittest
 
-import torch
-
-from e3nn import rs, o3
-from e3nn.tensor_product import ElementwiseTensorProduct, TensorProduct, TensorSquare
-
 
 class Tests(unittest.TestCase):
-
-    def test_elementwise_tensor_product(self):
-        torch.set_default_dtype(torch.float64)
-
-        Rs_1 = [(3, 0), (2, 1), (5, 2)]
-        Rs_2 = [(1, 0), (2, 1), (2, 2), (2, 0), (2, 1), (1, 2)]
-
-        Rs_out, m = rs.elementwise_tensor_product(Rs_1, Rs_2)
-        mul = ElementwiseTensorProduct(Rs_1, Rs_2)
-
-        x1 = torch.randn(1, rs.dim(Rs_1))
-        x2 = torch.randn(1, rs.dim(Rs_2))
-
-        y1 = mul(x1, x2)
-        y2 = torch.einsum('zi,zj->ijz', x1, x2)
-        y2 = m @ y2.reshape(-1, y2.shape[2])
-        y2 = y2.T
-
-        self.assertEqual(rs.dim(Rs_out), y1.shape[1])
-        self.assertLess((y1 - y2).abs().max(), 1e-7 * y1.abs().max())
-
-    def test_tensor_product(self):
-        torch.set_default_dtype(torch.float64)
-
-        Rs_1 = [(3, 0), (2, 1), (5, 2)]
-        Rs_2 = [(1, 0), (2, 1), (2, 2), (2, 0), (2, 1), (1, 2)]
-
-        Rs_out, m = rs.tensor_product(Rs_1, Rs_2, o3.selection_rule, sorted=True)
-        mul = TensorProduct(Rs_1, Rs_2, o3.selection_rule)
-
-        x1 = rs.randn(1, Rs_1)
-        x2 = rs.randn(1, Rs_2)
-
-        y1 = mul(x1, x2)
-        y2 = torch.einsum('zi,zj->ijz', x1, x2)
-        y2 = (m @ y2.reshape(rs.dim(Rs_1) * rs.dim(Rs_2), -1)).T
-
-        self.assertEqual(rs.dim(Rs_out), y1.shape[1])
-        self.assertLess((y1 - y2).abs().max(), 1e-7 * y1.abs().max())
-
-    def test_tensor_product_norm(self):
-        for Rs_in1, Rs_in2 in [([(1, 0)], [(2, 0)]), ([(3, 1), (2, 2)], [(2, 0), (1, 1), (1, 3)])]:
-            with o3.torch_default_dtype(torch.float64):
-                Rs_out, Q = rs.tensor_product(Rs_in1, Rs_in2, o3.selection_rule)
-
-                n = rs.dim(Rs_out)
-                I = torch.eye(n, dtype=Q.dtype())
-
-                d = ((Q @ Q.t()).to_dense() - I).pow(2).mean().sqrt()
-                self.assertLess(d, 1e-10)
-
-                d = ((Q.t() @ Q).to_dense() - I).pow(2).mean().sqrt()
-                self.assertLess(d, 1e-10)
-
-    def test_tensor_square_equivariance(self):
-        with o3.torch_default_dtype(torch.float64):
-            Rs_in = [(3, 0), (2, 1), (5, 2)]
-
-            sq = TensorSquare(Rs_in, o3.selection_rule)
-
-            x = rs.randn(Rs_in)
-
-            abc = o3.rand_angles()
-            D_in = rs.rep(Rs_in, *abc)
-            D_out = rs.rep(sq.Rs_out, *abc)
-
-            y1 = sq(D_in @ x)
-            y2 = D_out @ sq(x)
-
-            self.assertLess((y1 - y2).abs().max(), 1e-7 * y1.abs().max())
-
-    def test_tensor_square_norm(self):
-        for Rs_in in [[(1, 0), (1, 1)]]:
-            with o3.torch_default_dtype(torch.float64):
-                Rs_out, Q = rs.tensor_square(Rs_in, o3.selection_rule, normalization='component', sorted=True)
-
-                I1 = (Q @ Q.t()).to_dense()
-                I2 = torch.eye(rs.dim(Rs_out))
-
-                d = (I1 - I2).pow(2).mean().sqrt()
-                self.assertLess(d, 1e-10)
-
-    def test_tensor_product_in_out_norm(self):
-        for Rs_in1, Rs_out in [([(1, 0)], [(2, 0)]), ([(3, 1), (2, 2)], [(2, 0), (1, 1), (1, 3)])]:
-            with o3.torch_default_dtype(torch.float64):
-                Rs_in2, Q = rs.tensor_product(Rs_in1, o3.selection_rule, Rs_out)
-
-                n = rs.dim(Rs_out)
-                I = torch.eye(n, dtype=Q.dtype())
-
-                x = Q @ Q.t()
-                x = x.to_dense()
-                d = (x - I).pow(2).mean().sqrt()
-                self.assertLess(d, 1e-10)
+    pass
 
 
 if __name__ == '__main__':
